requirements.txt
discord.py>=2.3.0
yt-dlp>=2023.0.0
pynacl>=1.5.0
aiohttp>=3.8.0
requests>=2.0.0
ffmpeg-python
